% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_list.R
\name{data_to_list}
\alias{data_to_list}
\title{data_to_list}
\usage{
data_to_list(list_by)
}
\arguments{
\item{list_by}{pattern to search in actual directory for files to put in the list}
}
\description{
Takes all files which names match pattern provided for the user and put them in a list
}
\examples{
list_data_sigma <- data_to_list("K")
}
