% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_fisher.R
\name{enrichment_fisher}
\alias{enrichment_fisher}
\title{evaluates KO enrichment using the fisher test}
\usage{
enrichment_fisher(
  x,
  column_names,
  replace_by = NULL,
  phylogeny = selected_phylogeny
)
}
\arguments{
\item{x}{data frame where the test will be applied.}

\item{column_names}{Vector with column names to include in the test}

\item{replace_by}{Replace the Inf values in the odd.ratio matrix by a value.}

\item{phylogeny}{Vector of phylogenetic groups or rownames to include in the test}
}
\value{
a list with two matrixes: p.value and odd.ratio.
}
\description{
Takes a \code{data.frame} and a vector with column names
contained in the data.frame. Performs the fisher test to evaluate enrichement per KO
per phylogenetic group from selected phylogeny.
}
\examples{
my_data <- erba::sigma_filtered
fisher_evaluation <- enrichment_fisher(sigma_filtered, kos_sigma)
fisher_evaluation <- enrichment_fisher(sigma_filtered, kos_sigma, replace_by  = 100)
}
