% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_heatmap.R
\name{enrichment_heatmap}
\alias{enrichment_heatmap}
\title{make enrichment heatmap}
\usage{
enrichment_heatmap(x, tiff_name, my.limit, breaks)
}
\arguments{
\item{x}{odd.matrix from \link{enrichment_fisher}}

\item{tiff_name}{file name with .tiff extension}

\item{my.limit}{limits for color gradient}

\item{breaks}{a vector with breaks for color gradient}
}
\value{
heatmap exported as tiff file. Odd.ratio values are transformed as log(odd.ratio).
}
\description{
Takes the odd.ratio matrix from \link{enrichment_fisher} test. Odd.ratio values are plotted as log(odd.ratio).
}
\examples{
fisher_sigma <- erba::enrichment_fisher(sigma_filtered, kos_sigma, replace_by = 100)
my.odd.matrix <- fisher_sigma[[2]]
enrichment_heatmap(my.odd.matrix, "heatmap.tiff", my.limit = c(-10,10), breaks = seq(from = -10, to = 10, by = 2))
}
