% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_heatmap_dendrogram.R
\name{enrichment_heatmap_dendrogram}
\alias{enrichment_heatmap_dendrogram}
\title{Makes heatmap with dendrogram clustering}
\usage{
enrichment_heatmap_dendrogram(x, file.1, file.2, my.limits)
}
\arguments{
\item{x}{A matrix object with numeric values}

\item{file.1}{file name .png for image with default size}

\item{file.2}{file name .png for image with augmented size}

\item{my.limits}{Vector of two values with limits for color scale}
}
\value{
Two .png files with heatmap and dendrogram clustering
}
\description{
Makes heatmap with dendrogram clustering
}
\examples{
enrichment_heatmap_dendrogram(fisher_sigma[[2]], "sigma1.png", "sigma2.png", c(0,100))
}
