% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_heatmap_dendrogram2.R
\name{enrichment_heatmap_dendrogram2}
\alias{enrichment_heatmap_dendrogram2}
\title{Makes heatmap with dendrogram clustering}
\usage{
enrichment_heatmap_dendrogram2(
  x,
  file,
  my.limits,
  fill_by_log = TRUE,
  return.x = FALSE,
  return.y = FALSE,
  col.order = FALSE,
  row.order = FALSE
)
}
\arguments{
\item{x}{A matrix object with numeric values}

\item{file}{output file name}

\item{my.limits}{Vector of two values with limits for color scale}

\item{fill_by_log}{Logical value indicating whether heatmap is filled by a new value given by: enrichment value + 1e-5.}

\item{return.x}{Logical value indicating whether the function returns axis.x clustering}

\item{return.y}{Logical value indicating whether the function returns axis.y clustering}
}
\value{
A .png file with heatmap and dendrogram clustering. If return.x/return.y = TRUE, returns a list with clustering data.
}
\description{
Makes heatmap with dendrogram clustering
}
\examples{
enrichment_heatmap_dendrogram2(fisher_sigma[[2]], file = "sigma.png", my.limits  = c(-3,3))
enrichment_heatmap_dendrogram2((fisher_sigma[[2]] + 1e-3), file = "sigma.png", my.limits  = c(-3,3), fill_by_log = FALSE)
}
