% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erba.R
\docType{package}
\name{erba}
\alias{erba}
\title{erba: Expression Regulators from Bacteria and Archaea}
\description{
erba package includes functions to analyze and plot distribution of sigma factors,
transcription regulators and riboswitches either transcriptional and translational terminators.
}
\section{Maintainer}{
 Joselyn Chavez joselynchavezf@gmail.com
}

\section{See Also}{

Report bugs at https://github.com/josschavezf/erba/issues
}

\author{
M. Sc. Joselyn Chavez
}
