% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_specie2.R
\name{filter_specie2}
\alias{filter_specie2}
\title{Select organisms with the bigger genome according with a defined vector}
\usage{
filter_specie2(x, names = representative_organisms)
}
\arguments{
\item{x}{A data.frame}

\item{names}{A vector with organism names to select}
}
\description{
Select organisms with the bigger genome according with a defined vector
}
\examples{
# function applyied to one data frame
data <- filter_specie(list_data_sigma[[1]])
org_names <- data$V1
filter_specie2(data, names = org_names)
}
