% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kos.R
\name{get_kos}
\alias{get_kos}
\title{get_kos}
\usage{
get_kos(regulator)
}
\arguments{
\item{regulator}{Class of regulator: "sigma" for sigma factors or "tf" for transcription factors}
}
\description{
Imports KO names from local file
}
