% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_functions.R
\name{get_linear_coefficients}
\alias{get_linear_coefficients}
\title{get_linear_coefficients}
\usage{
get_linear_coefficients(inputTable)
}
\arguments{
\item{inputTable}{data frame to calculate regression model coefficients}
}
\description{
Get intercept and slope from the data's linear model
}
