% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_functions.R
\name{get_slopePerPhylum}
\alias{get_slopePerPhylum}
\title{Get slope from linear model per phylum}
\usage{
get_slopePerPhylum(data, x, y)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{x}{x-axis name}

\item{y}{y-axis name}
}
\value{
A \code{tibble} with m = slope for each phylum
}
\description{
Get slope from linear model per phylum
}
\examples{
get_slopePerPhylum(total_cogs,x =  "ORFs(X100)" , y ="Transcription Factors")
}
