% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_to_table.R
\name{group_to_table}
\alias{group_to_table}
\title{group_to_table}
\usage{
group_to_table(inputList)
}
\arguments{
\item{inputList}{list of data frames with same size and data order}
}
\description{
takes a list with data frames of similar size and put them in a single table.
The new table contains the regulator factors, organism key, total of regulators per organism, ORFs
per genome and phylogenetic group.
}
