% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_functions.R
\name{intercept}
\alias{intercept}
\title{intercept}
\usage{
intercept(x, y, m)
}
\arguments{
\item{x}{x axis}

\item{y}{y axis}

\item{m}{slope of distribution}
}
\description{
Get intercept value of points distribution with linear regression model
}
