% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny_percent.R
\name{phylogeny_percent}
\alias{phylogeny_percent}
\title{phylogeny_percent}
\usage{
phylogeny_percent(x, percent = 0)
}
\arguments{
\item{x}{A data.frame}

\item{percent_num}{cutoff of abundance percent}
}
\description{
get abundance percent per phylogenetic group and gives the groups above of threshold
}
\examples{
 # see phylogenies with representativity above 2\%
 major_phylogenies <- phylogeny_percent(table_sigma, 2)
}
