% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot_distribution}
\usage{
plot_distribution(data, filename, title)
}
\arguments{
\item{data}{A data.frame object}

\item{filename}{file name inside " "}

\item{title}{file name with .tiff extension}
}
\description{
plot total organisms versus ORFs per genome range
}
\examples{
plot_distribution(sigma_filtered,
filename = "bar_example.tiff",
title = "bar distribution")
}
