% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exception.R
\name{plot_exception}
\alias{plot_exception}
\title{plot_exception}
\usage{
plot_exception(
  x,
  filename = "figure.tiff",
  title = "",
  ylab = "",
  ymin = 0,
  ymax = 120,
  exception_group = NULL
)
}
\arguments{
\item{x}{A data.frame object}

\item{filename}{file name with .tiff extension}

\item{title}{plot name inside " "}

\item{ylab}{ylab name inside " "}

\item{ymin}{A number setting the ylab inferior limit.}

\item{ymax}{A number setting the ylab superior limit.}

\item{exception_group}{phylum to be used as exception}
}
\description{
plot total regulators versus ORFs per genome with double distribution
}
\examples{
plot_exception(data_riboswitch, filename = "riboswitch.tiff",
                title = "Riboswitches",
                ylab = "Riboswitches per genome",
                ymax = 80,
                exception_group = "Firmicutes" )
}
