% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{plot frequencies from vector}
\usage{
plot_freq(x, filename, title)
}
\arguments{
\item{x}{A vector}

\item{filename}{filename that ends with .tiff}

\item{title}{title for the plot}
}
\value{
A barplot into a tiff file.
}
\description{
takes a vector, counts frequency of the elements and makes a barplot.
}
\examples{
plot_freq(proteobacteria_labels, tiff_file = "proteobacteria_freq.tiff", plot_title = "Proteobacteria")
}
