% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{plot_lines}
\usage{
plot_lines(
  data,
  type = "general",
  filename = "figure.tiff",
  title = "",
  ylab = "",
  ymin = 0,
  ymax = 200
)
}
\arguments{
\item{data}{A data.frame object}

\item{type}{either "general" or "groups" for plotting per phylogenetic group}

\item{filename}{file name with .tiff extension}

\item{title}{plot name inside " "}

\item{ylab}{yLab name inside " "}

\item{ymin}{A number setting the ylab inferior limit.}

\item{ymax}{A number setting the ylab superior limit.}
}
\description{
plot total regulators versus ORFs per genome
}
\examples{
plot_lines(data = data_kos_sigma, type = "general", filename = "example_general.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ymax = 120)
plot_lines(data = data_kos_sigma, type = "groups", filename = "example_groups.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ymax = 120 )
}
