% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{plot_points}
\usage{
plot_points(
  data,
  type = "general",
  x,
  y,
  filename = "figure.tiff",
  title = "",
  xlab = "",
  ylab = "",
  ymax = 200,
  model.degree = 1
)
}
\arguments{
\item{data}{A data.frame object}

\item{type}{either "general" or "groups" for plotting per phylogenetic group}

\item{x}{variable name}

\item{y}{variable name}

\item{filename}{file name with .tiff extension}

\item{title}{plot title}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{ymax}{A number setting the ylab superior limit.}

\item{model.degree}{Degrees for the polynomial regression model. By default model.degree = 1.}
}
\description{
plot total regulators versus ORFs per genome
}
\examples{
plot_points(data = data_kos_sigma, type = "general", filename = "example_general.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ymax = 120)
plot_points(data = data_kos_sigma, type = "groups", filename = "example_groups.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ymax = 120 )
}
