% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_abundant_kos.R
\name{plot_principal}
\alias{plot_principal}
\title{plot_principal}
\usage{
plot_principal(vector_kos, data_frame, tiff_file, myTitle)
}
\arguments{
\item{vector_kos}{takes a vector with the column names to graph}

\item{data_frame}{data.frame object with the data}

\item{tiff_file}{name of the image written inside " "}

\item{myTitle}{graph title written inside " "}
}
\description{
Takes a vector of columns names,
obtains the total values per column and normalizes them
versus the total of regulators found per phylogenetic group.
Then plots in facet wrap style all the columns with the highest percent values.
Exports the graph in tiff format in the actual directory.
}
\examples{
plot_principal(my_principal_kos,sigma_filtered, "example.tiff", "Principal sigma factors")
}
