% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regulators.R
\name{plot_regulators}
\alias{plot_regulators}
\title{Plot TFs, Sigmas and Riboswitches versus genome size}
\usage{
plot_regulators(x, filename = "", title = "")
}
\arguments{
\item{x}{A \code{data.frame}}

\item{filename}{Filename}

\item{title}{Plot title}
}
\value{
A title.tiff file
}
\description{
Plot TFs, Sigmas and Riboswitches versus genome size
}
\examples{
data \%>\%
filter(phylum == "Actinobacteria") \%>\%
plot_regulators("Actinobacteria")
}
