% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_repressor.R
\name{plot_repressor}
\alias{plot_repressor}
\title{plot_repressor}
\usage{
plot_repressor(data, filename, title, ylab, ymin = 0, ymax = 150)
}
\arguments{
\item{data}{A data.frame object, needs to have the columns repressor and non_repressor for transcription factors}

\item{filename}{file name with .tiff extension}

\item{title}{plot name inside " "}

\item{ylab}{ylab name inside " "}

\item{ymin}{ylim min}

\item{ymax}{ylim max}
}
\description{
plot repressor and non_repressor transcription factors versus ORFs per genome
}
\examples{
plot_repressor(tf_repressor_non_repressor,
filename = "tf_repressor.tiff",
ylab = "Transcription factors",
title = "Transcription factors per genome",
ylim(0,120))
}
