% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_abundant_kos.R
\name{principal_ko}
\alias{principal_ko}
\title{principal_ko}
\usage{
principal_ko(data_frame, column, filter, from, to, k)
}
\arguments{
\item{data_frame}{data.frame object from where the data is taken}

\item{column}{column name where the filter is applied}

\item{filter}{pattern to match in the filtering column}

\item{from}{from which column the filter is doing}

\item{to}{last column to filter}

\item{k}{number of top columns you want to obtain}
}
\description{
prints the columns with highest total values from data frame
}
\examples{
principal_ko(sigma_filtered, sigma_filtered$group, "Proteobacteria", 1, 20, 6)
}
