% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_phylogeny.R
\docType{data}
\name{selected_phylogeny}
\alias{selected_phylogeny}
\title{selected_phylogeny}
\format{An object of class \code{character} of length 10.}
\usage{
selected_phylogeny
}
\description{
imports the phylogeny vector names with abundance above of 2%
}
\keyword{datasets}
