% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_riboswitch.R
\name{transform_riboswitch}
\alias{transform_riboswitch}
\title{transform_riboswitch}
\usage{
transform_riboswitch(x)
}
\arguments{
\item{x}{data frame with organism, gene and riboswitch columns.}
}
\value{
A \code{data.frame} with counts of riboswitches per organism and genome information.
}
\description{
re-order counts of riboswitches per organism and adds genomes information.
}
