% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_list.R
\name{data_to_list}
\alias{data_to_list}
\title{data_to_list}
\usage{
data_to_list(path = "./", pattern = NULL)
}
\arguments{
\item{path}{path to directory with files}

\item{pattern}{pattern to select files}
}
\description{
Takes all files which names match pattern provided for the user and put them in a list
}
\examples{
list_data_sigma <- data_to_list(path = "Documents/", pattern = "K")
}
