% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_functions.R
\name{get_correlation}
\alias{get_correlation}
\title{get_correlation}
\usage{
get_correlation(inputTable)
}
\arguments{
\item{inputTable}{table object}
}
\description{
Get correlation index of regulator's total number versus ORFs number per
organism. Shows correlation for each phylogenetic group.
}
