% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exception.R
\name{plot_exception}
\alias{plot_exception}
\title{plot_exception}
\usage{
plot_exception(data, filename, title, ylab, ylim, exception_group)
}
\arguments{
\item{data}{A data.frame object}

\item{filename}{file name with .tiff extension}

\item{title}{plot name inside " "}

\item{ylab}{ylab name inside " "}

\item{ylim}{ylab limit in format (x,y)}

\item{exception_group}{phylum to be used as exception}
}
\description{
plot total regulators versus ORFs per genome with double distribution
}
\examples{
plot_exception(data_riboswitch, filename = "riboswitch.tiff",
title = "Riboswitches",
ylab = "Riboswitches per genome",
ylim = ylim(0,120),
exception_group = "Firmicutes" )
}
