% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{plot_points}
\usage{
plot_points(
  data,
  type = "general",
  filename = "figure.tiff",
  title = "",
  ylab = "",
  ylim = ylim(0, 200)
)
}
\arguments{
\item{data}{A data.frame object}

\item{type}{either "general" or "groups" for plotting per phylogenetic group}

\item{filename}{file name with .tiff extension}

\item{title}{plot name inside " "}

\item{ylab}{yLab name inside " "}

\item{ylim}{yLab limit in format (x,y)}
}
\description{
plot total regulators versus ORFs per genome
}
\examples{
plot_points(data = data_kos_sigma, type = "general", filename = "example_general.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ylim = ylim(0,120))
plot_points(data = data_kos_sigma, type = "groups", filename = "example_groups.tiff", title = "Sigma factors", ylab = "Sigma factors per genome", ylim = ylim(0,120) )
}
